feat: simplify simple tree
